requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
distro==1.9.0
Django==5.1.6
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.3
openai==1.64.0
packaging==24.2
pillow==11.1.0
preshed==3.0.9
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.3
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.4
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2